+++ src/main/java/com/toptop/repository/CompanyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CompanyRepository extends JpaRepository<Company, Long>{
 
-    Company findOneByCompanyCod(int companyCod);
+    Company findOneByCompanyCod(Long companyCod);
 }
